Migration 0022: 30*24*3600*1000 overflows int4 — cast to bigint
The 7-to-30-day bump pushed the grace-window product to 2 592 000 000,
past Postgres's 4-byte int max, so the migration errored on every fresh
database (every test file died at setup). Applied DBs are unaffected;
they ran the 7-day version and keep that grace cap.

+++ server/migrations/0022_password_auth.sql
@@ -4,6 +4,8 @@
 -- permanent links get one grace redemption within 30 days.
 ALTER TABLE users ADD COLUMN IF NOT EXISTS password_hash TEXT;
 
+-- 30::bigint — as plain int4 literals the product (2 592 000 000) overflows
+-- before it ever meets the bigint, and the whole migration errors out.
 UPDATE invites SET permanent = FALSE,
-  expires_at = LEAST(expires_at, (extract(epoch from now()) * 1000)::bigint + 30*24*3600*1000)
+  expires_at = LEAST(expires_at, (extract(epoch from now()) * 1000)::bigint + 30::bigint*24*3600*1000)
 WHERE permanent;
